2nd Attempt - add checks in controllers

+++ api/src/main/java/api/controllers/users/UserController.java
@@ -24,6 +24,7 @@ import api.dtos.ErrorDto;
 import api.dtos.UserDto;
 import api.entities.User;
 import api.mapper.UserMapper;
+import api.services.RoleService;
 import api.services.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
@@ -42,6 +43,8 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 public class UserController {
     @Autowired
     private UserService userService;
+    @Autowired
+    private RoleService roleService;
     @Autowired
     private UserMapper userMapper;
     @Autowired
@@ -175,6 +178,11 @@ public class UserController {
     // endregion
     @DeleteMapping("/me")
     public void deleteMe(@AuthenticationPrincipal User user) {
+        if (user.getRoles().contains(roleService.get(RoleService.ADMIN_ROLE_NAME))
+            && userService.getAllContainingRole(roleService.get(RoleService.ADMIN_ROLE_NAME)).size() == 1
+        ) {
+            throw new IllegalStateException("Last admin user cannot be deleted.");
+        }
         userService.delete(user);
     }
 
@@ -354,6 +362,11 @@ public class UserController {
     @DeleteMapping("/{id}")
     @PreAuthorize("hasAuthority(@DbSetup.USER_WRITE)")
     public void deleteUser(@PathVariable int id) {
+        if (userService.getRoles(id).contains(roleService.get(RoleService.ADMIN_ROLE_NAME))
+            && userService.getAllContainingRole(roleService.get(RoleService.ADMIN_ROLE_NAME)).size() == 1
+        ) {
+            throw new IllegalStateException("Last admin user cannot be deleted.");
+        }
         userService.delete(id);
     }
 }
